Initial implementation for a CSV backed flat file map
file_map.go - Defines the API and interface for how to open and use the flat-file-map's
csv_map.go - Implementation using a CSV format.  Depending on the value type different encoding formats may be used, enabling a small but flat format.  Struct types are encoded using json.  When multiple values of the same struct are used, the format minimizes the output through a header record which defines the struct value names.  Values are encoded as a json array, primarily as a means of encoding the types while keeping code uniform.

Although JSON is simple, it is not efficent in a CSV due to the need to escape double quotes ("").  So this format will likely change soon.