Undo field type optimization from last two commits
This keeps (and expands) the testing changes, which also includes the switch to a `encodeValue` and `decodeValue` function for isolated testing.
However the field optimization is being undone.  After considering the previous fix further, there is the potential for more field variations between value instances of the same struct.  Because of that, I now believe the json overhead is worth the advantage of having each record typing be row independent.