csv map: Encode field name CRC into struct id
This commit makes it so that the struct fields are considered as part of the struct id.
This is important because if the struct changes across multiple Commit() and loads, the map will try to de-duplicate the fields.  If the fields have changed this may not occur correctly.

This commit fixes https://github.com/go-analyze/flat-file-map/issues/3