License change from MPL 2.0 to MIT
This commit re-licenses the code under MIT. Because all contributions so far are done by me, this is a point where changing the license is easier.

MPL 2.0 was chosen because it helps strengthen open source by encouraging contributions and changes to remain public.

However `go-analyze/charts` is MIT, and we would like to have a single license for all projects under `go-analyze`. Also MIT is more standard in the go community.