fix: Multiple struct optimization when `omitempty` is defined on the struct
When multiple matching structs are stored in the map, prior to this change only the encoded fields of the first struct were inspected. Because of our use of json encoding, that means that if the struct is defined with `omitempty`, any fields in the first instance not filled in the first instance wont be encoded for all remaining instances.

This change does an extra Unmarshal of all instances to ensure that all filled in fields are included in our map. Although there is a slight performance impact from this, it is an important check for us to do.