feat: Minimize encoded size by removing default / zero values
This change is a further reduction in the encoded flat file map size.
This is done by removing any fields which would be initialized to their default value. We are attempting to be very careful that loading the value will still match the state as exactly as possible. This includes cases like having pointers to empty maps, or empty slices, which will be reconstructed with the same state.

Tests were added to cover this as there are a lot of corner conditions to consider, including pointers to default vlaues (which we want to maintain), or slices of default values, and the fact that we need to accept a wide range of structs from outside our package. The tests added here were confirmed to behave the same as the current `main` / released logic.

An alternative prototyped was to encode the struct, decode into `map[string]interface{}`, prune, then re-encode. However we ran into more type issues, and maintaining slices of default values was challenging. This implementation is also more efficent, though admittedly with significant complexity.