fix: Fail to encode optimized embedded structs
This change is a re-attempt of PR #13. That change was reverted, with the fix released as `v0.3.4`. This change re-introduces the attempted optimization with the identified regression fixed; closes #14.

The key to fixing this was an added check for fields to be anonymous embedded struct. When this condition is detected we flatten out the fields to match the default and prior behavior done by the json Marshal.

While working on this it uncovered a potential panic condition from a nil pointer to the `json.Marshaler` interface. This was also corrected.

Testing was expanded to cover this, and other conditions with less coverage. This is a challenging and dynamic problem, but one which has potentially significant benefits in minimizing the encoded file output.