fix: Encode default values for basic types
This PR is a follow up from PR #13 and the fix in #15.

It was discovered that there was an unexpected change in encoding when the same struct is being provided with some fields left as default in some instances, but filled in others. Prior to the field removal a basic type (for example `int`) that was left in it's default value (`0`). After the change to remove the fields it is now encoded as `null`.

This does not result in an issue in recreating the struct from the json, but it does result in a larger json definition. `null` is larger than the default `[]`, `0`, etc. It also makes reading the file more difficult.

For that reason this behavior is not desired. We fix this in the `Commit` action by tracking the field types for all the instances we see. This allows us to define a default value for these types, and maintain the prior behavior for any structs which have had zero values pruned.