fix: Stop stripping basic default values from maps
This fix continues the efforts from PR's: #13, #15, #16

In those changes we are stripping default value fields. This issue does not exhibit itself in a decoding issue, but rather a conciseness and readability issue of the file (similar to #16).

Prior to these changes `stripZeroFields` would remove the values from a map if they were the default types. We now retain basic types in the maps regardless of the values. Added testing helps demonstrate this specific condition.