fix: Ensure struct fields are zero if absent in file
Prior to this change if a struct field was added after the file was written, or removed for optimization, and the `Get` call was provided a struct with the fields already set, those fields would not be cleared out.

This fixes #18 by adding an extra reflect call to zero out for the type.