feat: Add `CreateMemoryMap` and consistent error types
This change adds a function for returning a memory backed MutableFFMap.  This is generally only useful for testing since there is no way to export, persist, or restore the state (outside of the Set calls).
Internally this provides an improvement as KeyValueCSV now just references the in-memory implementation, this also ensures that the test behavior matches the CSV behavior exactly.

Additionally in `file_map.go` we now define consistent `EncodingError`, `TypeMismatchError`, and `ValidationError` so that error responses can be more easily checked. These structs also contain additional metadata in case batched errors need to be examined.